requirements.txt
asgiref==3.7.2
astroid==3.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
coverage==7.4.0
defusedxml==0.7.1
dill==0.3.7
Django==5.0
django-nose==1.4.7
djangorestframework==3.14.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
lxml==5.0.1
mccabe==0.7.0
nose==1.3.7
packaging==23.2
platformdirs==4.1.0
pluggy==1.3.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pylint==3.0.3
pytest==7.4.4
pytest-django==4.7.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomlkit==0.12.3
urllib3==2.1.0
